Locust (#892)
* Using a JSON logging pattern for now, assumes all messages are JSON objects

* Moves v1 endpoint to another class to seperate v1 -vs- v2 logging

* Refactors logging for a few initial classes

* Refactors logging for a few additional classes

* Refactors logging on remaining classes

* Minor log level fix

* Minor log level fix

* Updates Swagger logging and vss v2 response logging

* Updates error logging for OID parsing

* Initial update of locust.io test, need to test a larger collection of certs

* Updates locust.io test, and updates config rolling log max-history
