+++ locust/README.md
@@ -1,0 +1,9 @@
+# Locust.io Test
+
+To execute the test:
+
+```TEXT
+$ locust
+```
+
+Open http://localhost:8089
