+++ locust/requirements.txt
@@ -1,0 +1,2 @@
+locust==2.37.7
+locust-plugins==4.7.0
